fix compression when using apex.amp (#420)
- use the correct ctx when decompressing data
- use tensor.copy_ if tensor.set_ fails. apex.amp monkey-patches
optimizer.zero_grad to set all grads to None. torch creates a new p.grad
for each parameter. calling p.grad.set_() causes an error:

RuntimeError: set_storage is not allowed on a Tensor created from .data or .detach().
If your intent is to change the metadata of a Tensor (such as sizes / strides / storage / storage_offset)
without autograd tracking the change, remove the .data / .detach() call and wrap the change in a `with torch.no_grad():` block.
For example, change:
    x.data.set_(y)
to:
    with torch.no_grad():
        x.set_(y)

Signed-off-by: yulu.jia <yulu.jia@bytedance.com>